internal/db/dqlite: Begin heartbeat on database open
The heartbeat function was previously part of the dqlite dialFunc, but
since the heartbeat uses the dqlite client helpers, which results in
more calls to the dialFunc, this basically creates an infinite loop.
With the heartbeat lock, this doesn't pose any functional issue but
I've noticed my CPU spiking to 600% during heartbeats, which is
certainly not good!

This commit instead adds a loopHeartbeat helper which calls the hearbeat
function once per second. This helper is called once the database is
opened.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>