internal/state: Add State interface
Replaces the old State struct with an InternalState struct, and instead
exposes a State interface implemented by InternalState.

The old functions supplied to InternalState are now prefixed with
Internal.

Hooks are now directly a field on InternalState as we don't need to
worry about them being unassigned by the upstream project.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>